Attempt to update QPE thm _alpha_l_summed_abs
Attempt to update QPE thm _alpha_l_summed_abs, eventually having to save this with some of the final nb cells commented out for the time being. A critical step involves the combining of two fraction factors into one but getting an error message that 'Sin.deduce_in_number_set()' not implemented for the ComplexNonZero set. Not clear why this was not a problem in earlier manifestations of this theorem proof.