Establish simple RealPos-to-Real closure theorem for Log(a, b)
Related Issue: #265. Establish an initial simple RealPos-to-Real closure theorem for Log(a, b), for a,b in RealPos and a != 1. A few related closure theorems should be easy to add here, allowing some more precise number set inclusions involving RealPos and RealNeg depending on the base 'a' and antilog 'b'. These minimal modifications implemented to allow support for proof of QPE-related theorem e_value_ge_two.