Fix in Lambda._canonical_version
Track and use "num_indep_internal_bound_vars" to ensure canonical dummy variables are not contested.  The previous approach could fail (I saw it happen) if the Lambda parameter happens to be the same as one of the "internal bound variables" -- it got that count wrong.  This way avoids that issue.