Update QPE axioms with n def, eps_in_interval, and modification of _t_req
Related Issues: #97. Update QPE axioms with definition for 'n' (represented by the Literal _n) to denote the desired number of bits of precision in estimating the phase phi,  acknowledging that epsilon (the probability of failure) is in the real interval (0, 1], and modify _t_req to an inequality instead of equality.