Add QPE axioms n >= 2 and n <= t for convenience
Related Issue: #97. Add QPE axioms n >= 2 and n <= t for convenience, where n is the desired number of bits of precision and t is the number of qbits in the first register. The n >= 2 then comes up in later QPE-related proofs, especially theorems involving the last few paragraphs/equations in N & C pg 224. We should revisit the n >= 2 assumption when we have time.