Add bounding theorems for Abs(a) when b < a < c or b <= a <= c
Related Issues: #97. Add bounding theorems for Abs(a) in the numbers/absolute_value package for when b < a < c or b <= a <= c, and thereby augment/extend some bounding theorems already present in the package for the special symmetrical cases when -c < a < c or -c <= a <= c:
strict_upper_bound_asym_interval
weak_upper_bound_asym_interval
The weak version is then used directly in the proof of the QPE theorem _modabs_in_full_domain_simp (committed separately).