Revert 'underscore in title' fix
Meant to address #272, this cure was worse than the disease.
It turns out, the titles are fine in html but not the notebooks.  Fixing
it in the notebooks messed up the html.  So revert for now - we'll figure
this out later.

This commit also adds commented-out code for putting all dummy variables
in a special __pv_it/dummy folder, but this isn't working yet.