Update Div.cancelation() to catch Exp object in numerator
Related Issues: #73, #97, #283. Update Div.cancelation() method to catch and effectively deal with simple cases of Exp objects in numerator. The update relies indirectly on recent updates (saved earlier) to the Exp.factorization() method. The upgraded Div.cancelation() method is then used in a couple of the QPE theorem proofs (saved separately) to save a number of interactive coding lines in each (see, for example, the proof of QPE theorem _success_prob_guarantee).