Simplifying fix in coefficient_and_remainder of mult.py
Don't need to deal with the 'Div' case which wasn't quite right anyway.  If it's in a canonical form, it won't have a 'Div' except unless it's a literal rational.  The real fix from before was to get the canonical form of the coefficient -- dealing with the 'Div' case was superfluous.