Multi-chain transitivity shortcut
A TransitiveRelation may implement a "relate_across_chain_transitively" method that can be used to apply a theorem that effects a chain of multiple transitivity applications all-on-on.  This is implemented for Equals using the 'four_chain_transitivity' theorem; a more general 'transitivitity_chain' theorem was attempted, but I was having problems instantiating it (see Issue #264).