Add print-feedback statement in judgment._qed() method
Add a printing feedback statement in judgment._qed() method to (ideally) provide some quicker feedback to user when %qed cell is entered (to help keep user from wondering if the qed cell is churning away due to an as-yet unreported error). With a little bit of testing, not clear this is working as desired.