Fix inconsistency in Neg._build_canonical_form() method
Fix inconsistency in Neg._build_canonical_form() method that was causing an error in one of our QPE theorem proofs (somewhere along the way we were ending up with a canonical form of -Expr vs. (-1) * Expr somewhere else. This should fix things so the std canonical form of Neg(Expr) is the (-1)*Expr.