Augment generic_methods w/multi_disassociation() method
Augment generic_methods with a multi_disassociation() method that takes multiple "location" indices (indicating expressions to disassociate) and repeatedly calls the original disassociation() method. This is the eventual alternative we settled on when trying to modify the original apply_disassociation_thm() to deal with multiple simultaneous disassociations (but we ran into trouble because of the structure of the arguments not allowing us to maintain backward compatibility).