Fix assoc bug in VecSum.factors_extraction() method
Fix association bug in the VecSum.factors_extraction() method, changing so we don't wrap more parentheses around a single factor and only "associate" multiple factors. This was causing a recent problem where the factorization process was producing something like (c)*Sum() instead of simply c*Sum() which then was not being recognized as the same as c*Sum() and disassociation was not handling well a disassociation of a single term. Left original "if _num_factored > 0" commented-out in-place temporarily as reminder.