Fix Mult disassoc bug giving c(xx).disassociate(1) = c(xx)
Calling disassociate(1) on the expression 3*(e*e) kept producing 3*(e*e) (i.e. with no apparent change). Apparently this was due to 3*(e*e) being taken to 3*e*e, then the two e's being identified as something that should be combined, to then go to 3*(e*e) and eventually to 3*e^2, but the 3*(e*e) was also being preserved. This commit helps with most of this, but also does NOT address the related issue of dealing effectively with nested Mults.