Modify Div.cancelation() to allow Neg() numerator to proceed
Modify the Div.cancelation() method to accommodate the case where the entire numerator is a Neg object. When the numerator was a Neg object, the initial checks in the method was not letting it proceed, despite the fact that the later factorization() call worked just fine on the Neg object. This fix is a little kludgy, and the entire set of if/and/or/not clauses leading to the first error message should probably be reconsidered and reorganized.