Establish unique_int_in_interval theorem in real_numbers package.
Establish the unique_int_in_interval theorem in numbers/number_sets/real_numbers package. The theorem essentially says that there is a unique integer in any real interval about an integer if the interval is small enough. The theorem is then utilized in the proof of the QPE package theorem _scaled_delta_not_eq_nonzeroInt.