Establish set of sine_linear_bound_by_arg_x thms in trig pkg
Establish a set of sine_linear_bound_by_arg_x theorems in the trigonometry package, analogous to the set of sine_linear_bound_x theorems already there but now detailing bounds on Sin(t) by the function y = t for various domains of t. These are then used in some Sin class methods (in a later commit) to deduce upper and lower bounds on a Sin(t) function depending on the domain for the argument t.