Establish proof of _alpha_l_mod_two_pow_t_evaluation QPE thm
Establish the proof of the _alpha_l_mod_two_pow_t_evaluation QPE theorem. This is an extension/modification of the _alpha_l_evaluation theorem now quantified over all integers ell but then subscripted with (ell mod 2^t). This provides a slighter easy pathway for dealing with eventual "best" measurements in the form of a rounding-based measurement value b_r.