Core updates for efficiency: no side-effects for readily provability checks
When calling Expression.proven or Expression.readily_provable, we
now disable side-effect automation.  These are supposed to be quick
checks.  Also moved the check if a side-effect had already been
performed from Judgment to Proof where we also make sure that
auto_simplify and replacements are disabled and adding
'conclude_automation' to the key, ensuring the side-effects derived
with the same Proof._derive_side_effects "key" will be consistent.