Fixed Conditional.simplification and implemented Lambda.simplification
Needed to assume the condition when simplifying the value of a
Conditional.  Implemented Lambda.simplification to simplify the
body.  Also, make sure we derive assumption side-effects (if enabled)
at the top of Lambda._apply.  In extract_param_replacements, check
for "Trivial length requirement" which doesn't need to be included
as a requirement.