Improve transitive relation automation: known + provable transitive pairs
Transitive relations are now readily provable if we can find a combination of a known relation and a readily provable relation that combine transitively to prove the desired relation.  Also, made a better error message when trying to call a ".._on_both_sides" method on a relation when the domain isn't known.