Update QPE thms w/two delta_ne_0-based thms
Update QPE theorems with two very basic theorems using the constraint that delta_b ≠ 0 instead of ell ≠ 0. These two basic theorems should then be able to be used to produce alternative versions of several less-basic theorems to make them also based on the delta_b ≠ 0 constraint instead of the ell ≠ 0 constraint and thus to make the set of precision-guarantee-related theorems slightly more consistent with the presentation in Nielsen & Chuang (2010).