Tweak Add.deduce_not_equal, replacing is_usable with readily_provable
Tweak the Add.deduce_not_equal() method, replacing an is_usable() call on a theorem with a readily_provable() call on one of the theorem's conditions. This was motivated by an error arising in the _failure_upper_bound QPE thm nb that arose when the presumptions were modified to include ALL of proveit (which paradoxically then caused an error where there was none before).