Re-establish and update _scaled_delta_b_round_in_interval QPE thm
Re-establish and update proof for the _scaled_delta_b_round_in_interval QPE theorem. This was mistakenly thought to not be needed anymore, but we need it for the _delta_b_in_interval QPE theorem, which is then used in the _delta_b_is_zero_or_non_int QPE theorem.