Implementing covervative extensions by definition
We are adding "definitions" as a new category for a theory package (in addition to axioms, theorems, etc).  In the "definitions" notebook, you add "defining properties" for defining Literals of that package (either a Literal created in the "common expressions" notebook or an "_operator_" of an Operation).  These defining properties derive from a definition existence proof.  That is, by proving that something exists that satisfies all of the defining properties, then we can use a Literal as a shorthand for it.