Fixed a problem when adding new axiom/theorem names
The "force a reload" in _TheoryStorage.set_special_expressions
was not working as intended because Theory.getattr was get called
and reloading the axiom/theorem names before setting new ones.
The problem was only revealed when doing a clean build, so I'm not
sure how long this bug has been around.