build_alt.py: Building definition existence proofs
Still need to update and consolidate build.py and build_alt.py.  The issue is that build.py no longer works in newer versions of ipython -- build_alt.py was created as something that does work in newer versions but without all of the features (specifically, it doesn't save notebooks that it executes).