Add & prove 2 SetNotEquiv-definition-related theorems
Add the following 2 SetEquiv-related theorems to the sets/equivalence pkg: existential_set_not_equiv and set_not_equiv_existential, representing the two directions of the logical equivalence SetNotEquiv(A, B) <=> Exists_{x}(x is in one of the sets but not the other). The set_not_equiv_existential proof nb also has a temporary tangential "addendum" at the end documenting and exploring some interesting errors that arise when the Exists.derive_negated_forall() method is called repeatedly on the same object.