Generalize exp_eq thm by removing (domain=Complex) constraint
Generalize exp_eq thm in the numbers/exponentiation package by removing domain=Complex constraint. This more general thm can probably replace the analogous theorems currently specialized for integers, rationals, and reals. The exp_eq_real theorem itself is utilized in the proof of the irrationality of sqrt(2) and in the proof of the physics/quantum/QPE/_best_guarantee theorem.