Augment Database class with retrieve_records_as_dictionaries() method
Augment Database class (in database.py) with a retrieve_records_as_dictionaries() method that retrieves database records in the usual way (i.e. given certain criteria and then attributes to return) but now returns each found records in the form of a dictionary, in part to provide easier post-retrieval access and processing. Still some testing to be done to confirm everything works as expected. Related Issue: #312 .