Replace empty conjunction with TRUE in Lambda map creation
I've forgotten some of the details here but I believe these changes were working to address an occasional error where two expressions were not being recognized as the same because there was an empty conjunction And() being used as a condition in one of the expressions and it wasn't being simplified away. This might need to be revisited.