Establish SetOfAllMembership & SetOfAllNonmembership classes
Establish SetOfAllMembership & SetOfAllNonmembership classes, along with related methods and some supporting theorems, and extensive testing in the logic/sets/comprehension demonstrations notebook. The motivation for these updates/changes is partly the need for a "set-of-all-membership-in-bool" theorem and related method when dealing more extensively with the SetOfAll class of set comprehensions (for example in some of the QEC proofs related to code distance conservation). NOTE: several of the SetOfAll and SetOfAllMembership class methods still return errors when the SetOfAll utilizes multiple instance variables combined with multiple explicit conditions, and this is still a bit of a mystery. A few cases within the new methods also also behave oddly in that they return an error when first called but then work fine on the second and subsequent calls. Also a bit of a mystery.