Update set_of_all.py and set_of_all_membership.py as dev continues
Update the SetOfAll (in logic/sets/comprehension/set_of_all.py) and SetOfAllMembership (in ... set_of_all_membership.py) class and methods in continuing development. Current methods are only mdoerately functional, with errors arising in a number of cases where the presence of multiple explicit conditions cause incompatibilities between components of the instantiated theorems and the expected or "natural" assumptions the user might construct or prove. Development and testing is ongoing.