Update demonstrations and testing nbs for logic/sets/comprehension
Update the demonstrations and testing notebooks for logic/sets/comprehension, primarily for work on the SetOfAll and SetOfAllMembership classes and related methods such as .conclude(). The variety of SetOfAll and SetOfAllMembership methods work on simple forms but continue to have problems with situations in which multiple explicit conditions exist, which then cause incompatible instantiations of theorems related to the simpification (or lack of simplification) of nested conjunctions. See related previous commit of the set_of_all.py and set_of_all_membership.py files.