Update error catches and expr preservation in decorated_prover() method
Update some error catches and expression preservation in decorated_prover() method, mostly in response to recent efforts to update and debug the SetOfAll and SetOfAllMembership classes and related methods. More specifically, these are a few changes we thought would help avoid some errors arising in (for example) the SetOfAllMembership.conclude() method in which multiple explicit conditions often leads to instantiated theorems with components that do not match the "expected" or "natural" assumptions or judgments one would use, and those errors in turn seem to arise because of difficulties in dealing with nested conjunctions that are either created when we don't want them or simplified away when we want to keep them.