Update classes/membership/not_in_class.py and related theorems
Update classes/membership/not_in_class.py with typo corrections: In several places, the instantiation of theorems used the wrong target for substitution (possibly copied over from some other code and never updated for the new theorems, or the theorems were subtly changed to use different variables and then the code was not updated). These were all cases where the instantiation tried a substitution for S but needed a substitution for C. Also updated the class/membership theorems with a double_negated_membership theorem analogous to the one in sets/membership so the NotInClass.conclude_negation() method (already in the NotInClass methods but utilizing a non-existing theorem) should now work.