Further adapt Or theorems to XOr package
Further adapt the disjunction (Or) theorems to the exclusive disjunction (XOr) theory package. Most of these translated just fine, with a few left out because the translation not yet clear (for example, the or_if_any theorem acknowledging that a disjunction is TRUE if any of the operands are TRUE, is more challenging for the XOr since the XOr is TRUE if exactly an odd number of the operands are TRUE and the rest are FALSE).