Add in_bool_def_xor and unfold_is_bool_xor theorems to booleans theorems
We already have in_bool_def and unfold_is_bool, essentially forms of (or precursors for) the "law of the excluded middle", but we now add versions using the XOr (instead of the inclusive disjunction Or):  the in_bool_def_xor and unfold_is_bool_xor. Now we two options for "unfolding" what it means for A to be Boolean: the strong version of (A) xor (Not(A)) or the weaker version of (A) or (Not(A)).