Finish translating and testing into XOr class the Or class methods
Finish translating the Or class methods in the XOr class and update the exclusive_disjunction demonstrations.ipynb notebook with simple tests of the XOr methods. Most of the methods are quite satisfactory. A notable exception is the XOr.shallow_simplification() method, which is fairly primitive compared to the Or version. The weakness is due in part to the absence of quantification-related theorems and methods in the XOr class. Further development should consider possible theorems and methods acknowledging that an XOr evaluates to TRUE if and only if exactly an odd number of its operands evaluate to TRUE and the remaining operands evaluate to FALSE.