Establish k_power_set theory package in logic/sets
Establish k_power_set theory package in logic/sets, for basic concepts and definitions involving the set [S]^{k} of all k-element subsets of a set S. The commit includes the initial class KPowerSet (named because it is analogous to the more general power set and because it has the standard notation [S]^{k}). This will replace the KElemSubsets class created earlier (without its own theory package) and due to be deleted in future commit(s).