Update VerticesMembership.as_defined() method
Update VerticesMembership.as_defined() method to acknowledge that it works only for cases in which the domain Vertices(G) has as operand an explicit graph object of the form Graph(V,E) with explicit vertex set V, passing up a NotImplementedError() otherwise. Tested in vertices demonstrations.ipynb and in a related endpoints theorem proof notebook.