Begin to establish IntegerEven number set
Begin to establish the IntegerEven number set (the IntegerEvenSet class and the related IntegerEvenMembership class) and begin filling in details, including the standard theorem about membership being Boolean, along with beginning testing in the integers/demonstrations.ipynb nb. The IntegerEvenSet is tucked away inside the integers package and integers.py.