Add simple even/odd theorems to numbers_sets/integers
Add some simple even/odd theorems to numbers_sets/integers: claiming that 2, 4, 6, 8 are each even (i.e. in IntegerEven), that 1, 3, 5, 7, 9 are each odd (i.e. in IntegerOdd), and that the negative of an even (or odd) is still an even (or odd). Also added the single-digit claims to the numbers.__initi__.py for automatic loading.