Add commuted version of a binary even/odd theorem
Add a commuted version of a binary even/odd theorem claiming that the sum of an odd integer and an even integer is an odd integer. This will then be used to simplify some code in the Add.deduce_in_number_set() method. Also updates the add_int_even_from_even_odd theorem to allow an empty (zero operands) Add, which is axiomatically defined as 0 which is then also even.