Augment IntegerEvenMembership.conclude() method with Add case
Augment the IntegerEvenMembership.conclude() method with a case for the element being an instance of Add, in which case we return the element's Add.deduce_in_number_set() method using the IntegerEven number set. Not sure if this is reasonable based on the way @wwitzel likes to think about these things.