Augment IntegerOddMembership.conclude() method with Neg() element case
Analogous to what we did in the IntegerEvenMembership.conclude() method, augment the IntegerOddMembership.conclude() method with a case for the element being an instance of Neg(a) with 'a' itself being odd, in which case we the element is also odd (i.e. the negation of an odd is an odd).