Shift vertices & edges sub-theory content into more general graphs folder; establish walks sub-theory pkg
Remove the individual vertices and edges sub-theory packages, shifting the vertices & edges sub-theory content into the more general graphs folder. This reorganization includes a transfer of the sub-theory packages' axioms and theorems into the axioms and theorems for graphs (often with necessary name changes), and updating the graphs/demonstrations.ipynb to include testing and demonstrations that were present in the sub-theory packages' demonstrations notebooks. This commit also includes the establishment of the graphs/walks sub-theory package (essentially an empty package right now, but to be used for walks, trails, paths, circuits, and cycles, all conceptualized in terms of sequences of vertices).