Update logic/sets/common.ipyn with missing expr each_y_in_S_in_bool
Update logic/sets/common.ipyn with missing expr each_y_in_S_in_bool, which was causing problems in essential builds. Not clear how or why this ended up missing or necessary in this branch, but doesn't hurt to include it. Might create a conflict later when merging with other branches, but the conflict should be easy to resolve.