Add Closed(), EdgeSequence(), EdgeSet(), EulerianTrails() classes to graphs/walks
Add the Closed(), EdgeSequence(), EdgeSet(), EulerianTrails() classes to graphs/walks, and remove the HasEulerianTrail(G) and HasEulerianCircuit(G) (shifting them to graphs.py in next commit). Closed(W) represents the claim that the walk W is closed (i.e. its endvertices are the same). EdgeSequence(W) and EdgeSet(W) represent the edges of a walk W either as the original sequence (which might contain repeated edges) or as a set. EulerianTrails(G) represents the set of all Eulerian trails in graph G.