Add UNDEF literal
Add UNDEF literal, a common expression labeled as 'undefined' in logic/sets/functions, most immediately for use in the expression of function-related expressions where the function is not defined for some set of inputs. This is then used in an update to the functions_def axiomatic definition of sets/functions and related theorems.