Add FiniteGraphsLiteral() & FiniteGraphsMembership() classes
Add FiniteGraphsLiteral() & FiniteGraphsMembership() classes to the graphs theory package in graphs.py. This is in anticipation of some possible problems involving cardinalities of vertex and edge sets if we're not careful to distinguish finite and infinite graphs. Future commit(s) will actually rectify some errors where we've assumed that the size of a graph was finite (for example, assuming the size of a graph G, or the number of edges in G, was a Natural number). This also establishes the FiniteGraphs label or string in the graphs/common notebook for use in imports.