Update partitions thms to correctly use Disjoint()
Update the basic logic/sets/partitions theorems to correctly use the Disjoint() predicate (in the previous instantiation, I thought I could call Disjoint(P) where P was a set of sets, but instead Disjoint() takes individual set arguments). For convenience in the future it might be useful to have an alternative like DisjointSets(S) that takes a single set of sets as an argument.