Add deg_is_natural theorem to graphs sub-theory pkg
Add to the graphs sub-theory package the deg_is_natural theorem stating that the degree of any vertex is a natural number. Would also LIKE to have that the degree is ODD XOR EVEN, but the evens and odds are currently developed in a separate branch, and the XOR is developed in its own branch as well.