Establish the deg_is_odd_xor_even thm in QEC pkg
Establish the deg_is_odd_xor_even theorem in the physics/quantum/QEC sub-theory package stating that the degree of a vertex is either even or odd. This is not the best ultimate location for this theorem but needs a branch in which we've combined the graphs, evens/odds, and Xor-related packages.