Fix hash_id bug in _core_/_theory_storage.py
Fix a bug in _core_/_theory_storage.py in which 'hash_id' is somehow being used before it is defined. This is a bug that was also fixed independently in the separate branch 331-theory-of-graphs-development back on 4/24/2025, but never absorbed into master. See email w/WW on 4/24/2025 with suggestion for how to fix.