expr.py: ProofFailure check and disprovability of ExprTuples
Check for ProofFailure when trying 'conclude_negation'.  Also, while an ExprTuple cannot be TRUE, it can now be disprovable (even if it cannot be FALSE) because negation is now defined as implying FALSE which can be TRUE without be a Boolean (it just means it isn't TRUE).  This is all consistent in Prove-It's logic where we don't bother to do type-checking.