Augment Exists().unfold() and Exists().definition() methods
Augment Exists().unfold() and Exists().definition() methods to also deal with the cases of Exists(x, Not(P(x))) instead of just the Exists(x, P(x)) cases, and establish some minimal testing of the methods in the existence demonstrations notebook. The  augmentations utilize the alternative existential unfolding, folding, and equality theorems in recent previous commits.