Update Exists().definition() to use more specific expr preservation approach
Update the Exists().definition() method to use a more specific expression-preservation approach instead of the 'preserve_all = True' approach. The 'preserve_all = True' leaves preserved the "empty" condition cases which then show up as a redundant 'T' in the eventual expression result; this more precise preservation method constructs the expected rhs expression and explicitly preserves that instead. This performs well in a handful of examples in the existence demonstrations notebook.