Add comprehension_superset_reduction thm to comprehension sub-theory pkg
Add the comprehension_superset_reduction theorem to the comprehension sub-theory package, which is anticipated to help with one or more proofs in the QEC sub-theory package. Basically, the theorem encapsulates the fact that, given that Q(x) does not hold for x in A, the `SetOfAll` {x | Q(X)}_{x in S} reduces to {x | Q(x)}_{x in (S-A)}. The theorem will be used in a SetOfAll() method to deduce the equality.